backend/requirements.txt
asgiref==3.11.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.8.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
fonttools==4.61.1
idna==3.11
Markdown==3.10
pillow==12.1.0
pycparser==2.23
pydyf==0.12.1
PyJWT==2.10.1
pymongo==4.16.0
pyphen==0.17.2
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
urllib3==2.6.2
weasyprint==67.0
webencodings==0.5.1
zopfli==0.4.0
